[feat] #9 필터 적용 기능 구성

+++ Filtee/Core/DTO/Filter/PhotoMetadataDTO.swift
@@ -10,7 +10,7 @@ import Foundation
 struct PhotoMetadataDTO: DTO {
     let camera: String?
     let lensInfo: String?
-    let focalLength: Int?
+    let focalLength: Double?
     let aperture: Double?
     let iso: Int?
     let shutterSpeed: String?
